feat: add health_check.host_header to customize livenessProbe HTTP Host header (#61)
The Host header is the "To:" field of an HTTP request. It indicates what
server the HTTP request was originally sent to.

Django has a setting called ALLOWED_HOSTS. It rejects requests that have
an unexpected Host header (e.g. the request was originally sent to
evil.site instead of edx.org).

Right now, Django is rejecting Kubernetes' HTTP GET livenessProbe
because it has an unexpected Host header.

This commit allows us to customize the Host header Kubernetes'
livenessProbe uses so the prove is not rejected.