fix: Add className variable ISRE-1476
It turns out we aren't able to create new ingresses that set
both the annotation and the ingressClass variable in the spec
so this changes it so we can specify the one we want and leave the
other out. This allows us to have older ingresses for the older
controller in parallel with newer style ingresses

Error looks like this
annotations.kubernetes.io/ingress.class: Invalid value: "internal-3":
can not be set when the class field is also set (retried 5 times).