fix: Make ingress names stable ISRE-1476
This will prevent ingresses and certs needing to be recreated or
updated when the ingresses array is updated. THe old system named
the ingresses and tls secrets based on the array index, so changing
the order of the array or removing items would uncesscarily change
ingresses and cause certs to be regenerated. The new stable names
only apply to newer ingresses that specify className via spec
instead of the deprecated annotation, so as to not disturb existing
ingresses.