fix: Prevent YAML injection in the worker command (#89)
Currently the worker command cannot contain the string `: ` because that
will be rendered directly into the YAML and be interpreted as a key/value
pair separator. This prevents the use of JSON in our Event Bus consumer
commands, required for event replay.

The solution is to use the `quote` function, which escapes all control
characters (and double-quotes), and then wraps the string in double-quotes.
However, YAML will not respect the escapes unless the double-quoted string
is the *entire* value, which it is not in this case. (The presence of
`source` etc. in front will tell YAML that this is a "plain flow scalar"
rather than a "double-quoted flow scalar.") So, we're only going to use
`quote` for its ability to transform control characters (critically,
including newlines, which might be in the inserted script). We'll use
the output as a quoted string *for bash*, and then convert the whole value
into a "stripping folded block scalar" multiline string so that colons are
not misinterpreted as key-value mapping separators.

This is still not perfect, as there is no escaping happening at the shell
script level, which is now required due to the addition of `bash -c`. But
this is only likely to come up with things like `!` and variables, which
are not likely to be in use at the moment.